back/requirements.txt
aio-pika==9.4.0
aioboto3==12.1.0
aiobotocore==2.8.0
aiofile==3.8.8
aiohttp==3.9.1
aioitertools==0.11.0
aiormq==6.8.0
aiosignal==1.3.1
alembic==1.13.1
amqp==5.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-property==0.2.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
backports.tarfile==1.2.0
billiard==4.2.1
boto3==1.33.1
botocore==1.33.1
bson==0.5.10
caio==0.9.13
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-log==0.4.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
deprecation==2.1.0
distlib==0.3.8
dnspython==2.4.2
docutils==0.21.2
dotty-dict==1.3.1
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fast-depends==2.3.1
fastapi==0.105.0
faststream==0.3.14
filelock==3.13.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.5.33
idna==3.6
importlib_metadata==8.5.0
iniconfig==2.0.0
invoke==1.7.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
jwcrypto==1.5.6
keyring==25.4.1
kombu==5.4.2
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
minio==7.2.0
miniopy-async==1.18
more-itertools==10.5.0
multidict==6.0.4
nh3==0.2.18
nodeenv==1.8.0
numpy==1.26.3
openpyxl==3.1.5
packaging==23.2
pamqp==3.3.0
pillow==10.2.0
pkginfo==1.11.1
platformdirs==4.1.0
pluggy==1.3.0
polars==0.20.4
pre-commit==3.6.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==3.17.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==2.10.1
python-jose==3.3.0
python-keycloak==5.1.1
python-multipart==0.0.6
pywin32-ctypes==0.2.3
PyYAML==6.0.1
readme_renderer==44.0
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rsa==4.9
s3transfer==0.8.0
semver==2.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
tomlkit==0.7.0
tqdm==4.66.1
twine==3.8.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2025.1
urllib3==1.26.18
uvicorn==0.27.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.2
